Migrando Api a Spring Boot 3.0

+++ src/main/java/com/pimentelprojects/api/controller/AppController.java
@@ -3,6 +3,8 @@ package com.pimentelprojects.api.controller;
 import com.pimentelprojects.api.models.Client;
 import com.pimentelprojects.api.service.ClientService;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 
@@ -10,30 +12,66 @@ import java.util.List;
 import java.util.Optional;
 
 @RestController
-@RequestMapping("/api/v1")
+@RequestMapping("/api/v1/clients")
 public class AppController {
 
-    @Autowired
+
     private ClientService clientService;
 
-    @GetMapping("/clients")
-    public List<Client> getAllClients(){
-        return clientService.findAll();
+    @Autowired
+    public AppController(ClientService clientService) {
+        this.clientService = clientService;
     }
 
-    @GetMapping("/clients/find/{id}")
-    public Optional<Client> getClientById(@PathVariable("id") Long id){
-        return clientService.findClientById(id);
+    @GetMapping
+    public ResponseEntity<List<Client>> getAllClients(){
+        if(clientService.findAll().isEmpty()){
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+        return new ResponseEntity<List<Client>>(clientService.findAll(), HttpStatus.OK);
     }
 
-    @PostMapping("clients/save")
-    public void saveClient(@RequestBody Client client){
+    @GetMapping("{id}")
+    public  ResponseEntity<Optional<Client>> getClientById(@PathVariable("id") Long id){
+        if(!clientService.existById(id)){
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+        return new ResponseEntity<Optional<Client>>(clientService.findClientById(id), HttpStatus.OK);
+    }
+
+    @PostMapping
+    public ResponseEntity<String> saveClient(@RequestBody Client client){
+        if(client.getId() != null){
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
+
         clientService.saveClient(client);
+
+        return new ResponseEntity<String>("Usuario creado exitosamente", HttpStatus.CREATED);
     }
 
-    @DeleteMapping("clients/delete/{id}")
-    public void deleteUser(@PathVariable("id") Long id){
+    @PutMapping
+    public ResponseEntity<String> updateClient(@RequestBody Client client){
+        if(client.getId() == null){
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
+
+        if(!clientService.existById(client.getId())){
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
+
+        clientService.saveClient(client);
+
+        return new ResponseEntity<String>("Usuario actualizado exitosamente", HttpStatus.OK);
+    }
+
+    @DeleteMapping("{id}")
+    public ResponseEntity<String> deleteUser(@PathVariable("id") Long id){
+        if(!clientService.existById(id)){
+            return new ResponseEntity<>(HttpStatus.BAD_REQUEST);
+        }
         clientService.deleteClientById(id);
+        return new ResponseEntity<String>("Usuario eliminado exitosamente", HttpStatus.OK);
     }
 
 }
